Refactor TTS API structure: modularize endpoints, update tests, and enhance coverage
