+++ tts/fishaudio/__init__.py
@@ -1,0 +1,1 @@
+# FishAudio TTS provider package
