+++ tts/fishaudio/v1/__init__.py
@@ -1,0 +1,1 @@
+# FishAudio TTS v1 API package
